Remove tryAquire() and non-tagged rate limiter
- tryAquire() can be achieved by passing 0 as timeout for the other
variants.
- non-tagged rate limiter is confusing because the interface supports
  both tagged and non-tagged. So removing it and implement a default tag
for the non-tagged operations.
