+++ samza-core/src/test/java/org/apache/samza/util/TestEmbeddedTaggedRateLimiter.java
@@ -18,13 +18,21 @@
  */
 package org.apache.samza.util;
 
+import java.lang.reflect.Field;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.apache.samza.SamzaException;
+import org.apache.samza.config.Config;
+import org.apache.samza.container.SamzaContainerContext;
+import org.apache.samza.task.TaskContext;
 import org.junit.Assert;
 import org.junit.Test;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 
 public class TestEmbeddedTaggedRateLimiter {
@@ -34,11 +42,70 @@ public class TestEmbeddedTaggedRateLimiter {
   final static private int TARGET_RATE_RED = 1000;
   final static private int TARGET_RATE_PER_TASK_RED = TARGET_RATE_RED / NUMBER_OF_TASKS;
   final static private int TARGET_RATE_GREEN = 2000;
-  final static private int TARGET_RATE_PER_TASK_GREEN = TARGET_RATE_GREEN / NUMBER_OF_TASKS;
   final static private int INCREMENT = 2;
+
+  final static private int TARGET_RATE = 4000;
+  final static private int TARGET_RATE_PER_TASK = TARGET_RATE / NUMBER_OF_TASKS;
 
   @Test
   public void testAcquire() {
+    RateLimiter rateLimiter = new EmbeddedTaggedRateLimiter(TARGET_RATE);
+    initRateLimiter(rateLimiter);
+
+    int count = 0;
+    long start = System.currentTimeMillis();
+    while (System.currentTimeMillis() - start < TEST_INTERVAL) {
+      rateLimiter.acquire(INCREMENT);
+      count += INCREMENT;
+    }
+
+    long rate = count * 1000 / TEST_INTERVAL;
+    verifyRate(rate);
+  }
+
+  @Test
+  public void testAcquireWithTimeout() {
+    RateLimiter rateLimiter = new EmbeddedTaggedRateLimiter(TARGET_RATE);
+    initRateLimiter(rateLimiter);
+
+    boolean hasSeenZeros = false;
+
+    int count = 0;
+    int callCount = 0;
+    long start = System.currentTimeMillis();
+    while (System.currentTimeMillis() - start < TEST_INTERVAL) {
+      ++callCount;
+      int availableCredits = rateLimiter.acquire(INCREMENT, 20, MILLISECONDS);
+      if (availableCredits <= 0) {
+        hasSeenZeros = true;
+      } else {
+        count += INCREMENT;
+      }
+    }
+
+    long rate = count * 1000 / TEST_INTERVAL;
+    verifyRate(rate);
+    junit.framework.Assert.assertTrue(Math.abs(callCount - TARGET_RATE_PER_TASK * TEST_INTERVAL / 1000 / INCREMENT) <= 2);
+    junit.framework.Assert.assertFalse(hasSeenZeros);
+  }
+
+  @Test(expected = IllegalStateException.class)
+  public void testFailsWhenUninitialized() {
+    new EmbeddedTaggedRateLimiter(100).acquire(1);
+  }
+
+  @Test(expected = IllegalArgumentException.class)
+  public void testFailsWhenUsingTags() {
+    RateLimiter rateLimiter = new EmbeddedTaggedRateLimiter(10);
+    initRateLimiter(rateLimiter);
+    Map<String, Integer> tagToCredits = new HashMap<>();
+    tagToCredits.put("red", 1);
+    tagToCredits.put("green", 1);
+    rateLimiter.acquire(tagToCredits);
+  }
+
+  @Test
+  public void testAcquireTagged() {
     RateLimiter rateLimiter = createRateLimiter();
 
     Map<String, Integer> tagToCount = new HashMap<>();
@@ -67,36 +134,7 @@ public class TestEmbeddedTaggedRateLimiter {
   }
 
   @Test
-  public void testTryAcquire() {
-
-    RateLimiter rateLimiter = createRateLimiter();
-
-    Map<String, Integer> tagToCount = new HashMap<>();
-    tagToCount.put("red", 0);
-    tagToCount.put("green", 0);
-
-    Map<String, Integer> tagToCredits = new HashMap<>();
-    tagToCredits.put("red", INCREMENT);
-    tagToCredits.put("green", INCREMENT);
-
-    long start = System.currentTimeMillis();
-    while (System.currentTimeMillis() - start < TEST_INTERVAL) {
-      Map<String, Integer> resultMap = rateLimiter.tryAcquire(tagToCredits);
-      tagToCount.put("red", tagToCount.get("red") + resultMap.get("red"));
-      tagToCount.put("green", tagToCount.get("green") + resultMap.get("green"));
-    }
-
-    {
-      long rate = tagToCount.get("red") * 1000 / TEST_INTERVAL;
-      verifyRate(rate, TARGET_RATE_PER_TASK_RED);
-    } {
-      long rate = tagToCount.get("green") * 1000 / TEST_INTERVAL;
-      verifyRate(rate, TARGET_RATE_PER_TASK_GREEN);
-    }
-  }
-
-  @Test
-  public void testAcquireWithTimeout() {
+  public void testAcquireWithTimeoutTagged() {
 
     RateLimiter rateLimiter = createRateLimiter();
 
@@ -126,7 +164,7 @@ public class TestEmbeddedTaggedRateLimiter {
   }
 
   @Test(expected = IllegalStateException.class)
-  public void testFailsWhenUninitialized() {
+  public void testFailsWhenUninitializedTagged() {
     Map<String, Integer> tagToTargetRateMap = new HashMap<>();
     tagToTargetRateMap.put("red", 1000);
     tagToTargetRateMap.put("green", 2000);
@@ -139,7 +177,7 @@ public class TestEmbeddedTaggedRateLimiter {
     tagToCredits.put("red", 1);
     tagToCredits.put("green", 1);
     RateLimiter rateLimiter = new EmbeddedTaggedRateLimiter(tagToCredits);
-    TestEmbeddedRateLimiter.initRateLimiter(rateLimiter);
+    initRateLimiter(rateLimiter);
     rateLimiter.acquire(1);
   }
 
@@ -154,8 +192,36 @@ public class TestEmbeddedTaggedRateLimiter {
     tagToTargetRateMap.put("red", TARGET_RATE_RED);
     tagToTargetRateMap.put("green", TARGET_RATE_GREEN);
     RateLimiter rateLimiter = new EmbeddedTaggedRateLimiter(tagToTargetRateMap);
-    TestEmbeddedRateLimiter.initRateLimiter(rateLimiter);
+    initRateLimiter(rateLimiter);
     return rateLimiter;
   }
 
+  private void verifyRate(long rate) {
+    // As the actual rate would likely not be exactly the same as target rate, the calculation below
+    // verifies the actual rate is within 5% of the target rate per task
+    junit.framework.Assert.assertTrue(Math.abs(rate - TARGET_RATE_PER_TASK) <= TARGET_RATE_PER_TASK * 5 / 100);
+  }
+
+  static void initRateLimiter(RateLimiter rateLimiter) {
+    Config config = mock(Config.class);
+    TaskContext taskContext = mock(TaskContext.class);
+    SamzaContainerContext containerContext = mockSamzaContainerContext();
+    when(taskContext.getSamzaContainerContext()).thenReturn(containerContext);
+    rateLimiter.init(config, taskContext);
+  }
+
+  static SamzaContainerContext mockSamzaContainerContext() {
+    try {
+      Collection<String> taskNames = mock(Collection.class);
+      when(taskNames.size()).thenReturn(NUMBER_OF_TASKS);
+      SamzaContainerContext containerContext = mock(SamzaContainerContext.class);
+      Field taskNamesField = SamzaContainerContext.class.getDeclaredField("taskNames");
+      taskNamesField.setAccessible(true);
+      taskNamesField.set(containerContext, taskNames);
+      taskNamesField.setAccessible(false);
+      return containerContext;
+    } catch (Exception ex) {
+      throw new SamzaException(ex);
+    }
+  }
 }
